added javadoc, made methods easier to read in the Loader and fixed checkstyle

+++ util.image/src/main/java/edu/kit/ifv/trafficspvisualizer/util/image/StandardImageGenerator.java
@@ -63,7 +63,7 @@ public class StandardImageGenerator extends ImageGenerator{
         this.attributes = attributes;
         this.dataObject = dataObject;
         this.situationIndex = situationIndex;
-        int attributeFontSize = (int) (height / 10);
+        int attributeFontSize = height / 10;
         this.attributeFont = new Font("Arial", Font.BOLD, attributeFontSize);
         javafx.scene.paint.Color fxColor = choiceOption.getColor();
         this.color = new java.awt.Color((float) fxColor.getRed(),
